Moved the component to its own folder
