+++ src/components/SearchingredientsBar/SearchIngredients.css
